Update to 32bit! No documentation yet...

+++ a3k-src/execute.h
@@ -13,8 +13,9 @@ extern "C" {
 	/*
 	 * Returns the current position (value) of the program counter.
 	 * Returns 0 if the program hasn't started.
+	 * 0 is also a valid running value.
 	 */
-	int current_pc();
+	uint32_t current_pc();
 
 	/*
 	 * Resets the PC to 0 for a new program.
@@ -26,9 +27,13 @@ extern "C" {
 	 * passed in as parameters. The PC is locally stored.
 	 * Returns 1 if the program is still active.
 	 * Returns 0 if the program has finished (END) or hasn't started.
+	 * Returns MEM_ERR for uninitialized memory or register files.
+	 * Returns OUT_OF_RANGE for out of range memory addresses.
+	 * Returns IO_ERR for syntax errors in the file encoding.
+	 * Returns DIVIDE_BY_ZERO for the obvious.
 	 * Please do not change any of the registers mid-execution, obviously.
 	 */
-	int execute_next(uint16_t *, int *);
+    int execute_next(uint32_t *, uint32_t *);
 	
 #ifdef __cplusplus
 }
